requirements.txt
optuna==4.4.0
numpy==2.2.5
scikit-learn==1.6.1
torch==2.7.0
pandas==2.2.3
matplotlib==3.10.3
umap-learn==0.5.7
seaborn==0.13.2